🏗️ Architect: Modularized planetary events calculations
Refactored the 400-line `apts/events/calculations/planetary.py` "God Module" into a modular package structure in `apts/events/calculations/planetary/`.

Logic is now split into five specialized sub-modules:
- `conjunctions.py`: Handles all types of planetary and lunar-planetary conjunctions.
- `extrema.py`: Handles observational extrema like greatest brilliancy, elongations, and stationary points.
- `orbits.py`: Handles orbital events such as oppositions and perihelion/aphelion.
- `alignments.py`: Handles multi-body alignments and occultations.
- `jovian.py`: Handles Jupiter-specific events including moon phenomena and mutual events.

The public API is preserved via re-exports in `__init__.py`. Unit test mocks in `tests/unit/test_events.py` have been updated to reflect the new module paths. Unused imports were removed and code style was improved based on internal review.

🎯 Benefit: Improves maintainability and adheres to the Single Responsibility Principle by decomposing a monolithic calculation module.
