--- apts/events/calculations/planetary.py
@@ -1,400 +1,0 @@
-import logging
-import time
-from datetime import timezone
-from itertools import combinations
-from concurrent.futures import as_completed
-
-from ... import skyfield_searches
-from ...utils import planetary
-from ..rarity import get_rarity
-from ..duration import get_duration
-
-logger = logging.getLogger(__name__)
-utc = timezone.utc
-
-def calculate_venus_greatest_brilliancy(observer, start_date, end_date):
-    start_time = time.time()
-    events = skyfield_searches.find_venus_greatest_brilliancy(
-        observer, start_date, end_date
-    )
-    for event in events:
-        event["rarity"] = get_rarity("Venus Greatest Brilliancy", event)
-        event["duration"] = get_duration("Venus Greatest Brilliancy", event)
-    logger.debug(
-        f"--- calculate_venus_greatest_brilliancy: {time.time() - start_time}s"
-    )
-    return events
-
-def calculate_conjunctions(ts, observer, start_date, end_date, executor):
-    start_time = time.time()
-    events = []
-    planets_data = {
-        p: planetary.get_skyfield_obj(p) for p in planetary.CONJUNCTION_PLANETS
-    }
-    moon = "moon"
-    moon_display_name = "Moon"
-    moon_obj = planetary.get_skyfield_obj(moon)
-
-    # Pre-compute positions for all bodies involved in conjunctions
-    step = 0.01
-    num_steps = int(
-        (end_date - start_date).total_seconds() / (step * 86400)
-    )
-    if num_steps < 2:
-        num_steps = 2
-    times = ts.linspace(
-        ts.utc(start_date), ts.utc(end_date), num_steps
-    )
-
-    precomputed = {}
-    obs_at_times = observer.at(times)
-
-    def _observe(name, obj):
-        return name.lower(), obs_at_times.observe(obj)
-
-    # Optimization: Parallelize high-precision observations using the passed executor.
-    # Benchmarking confirms ~2.2x speedup for this pre-computation step.
-    precompute_futures = [
-        executor.submit(_observe, p, obj) for p, obj in planets_data.items()
-    ]
-    precompute_futures.append(executor.submit(_observe, moon, moon_obj))
-
-    for future in as_completed(precompute_futures):
-        name, pos = future.result()
-        precomputed[name] = pos
-
-    futures = {}
-
-    for (p1, p1_obj), (p2, p2_obj) in combinations(planets_data.items(), 2):
-        p1_name = planetary.get_simple_name(p1)
-        p2_name = planetary.get_simple_name(p2)
-        future = executor.submit(
-            skyfield_searches.find_conjunctions_between_moving_bodies,
-            observer,
-            p1,
-            [(p2_name, p2_obj)],
-            start_date,
-            end_date,
-            threshold_degrees=5.0,
-            precomputed_positions=precomputed,
-        )
-        futures[future] = p1_name
-
-    planet_list = [
-        (planetary.get_simple_name(p), obj) for p, obj in planets_data.items()
-    ]
-    future_moon = executor.submit(
-        skyfield_searches.find_conjunctions_between_moving_bodies,
-        observer,
-        moon,
-        planet_list,
-        start_date,
-        end_date,
-        threshold_degrees=5.0,
-        precomputed_positions=precomputed,
-    )
-    futures[future_moon] = moon_display_name
-
-    for future in as_completed(futures):
-        obj1_name = futures[future]
-        found_events = future.result()
-        for event in found_events:
-            event["type"] = "Conjunction"
-            event["event"] = "Conjunction"
-            event["object1"] = obj1_name
-            event["rarity"] = get_rarity("Conjunction", event)
-            event["duration"] = get_duration("Conjunction", event)
-            events.append(event)
-    logger.debug(f"--- calculate_conjunctions: {time.time() - start_time}s")
-    return events
-
-def calculate_oppositions(observer, start_date, end_date, executor):
-    start_time = time.time()
-    events = []
-    planets = planetary.OPPOSITION_PLANETS
-    futures = [
-        executor.submit(
-            skyfield_searches.find_oppositions,
-            observer,
-            p,
-            start_date,
-            end_date,
-        )
-        for p in planets
-    ]
-    for future in as_completed(futures):
-        found_events = future.result()
-        for event in found_events:
-            event["type"] = "Opposition"
-            simple_name = planetary.get_simple_name(event["planet"])
-            event["event"] = "Opposition"
-            event["object"] = simple_name
-            del event["planet"]
-            event["rarity"] = get_rarity("Opposition", event)
-            event["duration"] = get_duration("Opposition", event)
-        events.extend(found_events)
-    logger.debug(f"--- calculate_oppositions: {time.time() - start_time}s")
-    return events
-
-def calculate_highest_altitudes(observer, start_date, end_date, executor):
-    start_time = time.time()
-    events = []
-    futures = {
-        executor.submit(
-            skyfield_searches.find_highest_altitude,
-            observer,
-            planetary.get_skyfield_obj(planet_name),
-            start_date,
-            end_date,
-        ): planet_name
-        for planet_name in ["mercury", "venus"]
-    }
-    for future in as_completed(futures):
-        planet_name = futures[future]
-        t, alt = future.result()
-        if t:
-            simple_name = planetary.get_simple_name(planet_name)
-            event_data = {
-                "date": t.astimezone(utc),  # type: ignore
-                "event": "Highest altitude",
-                "object": simple_name,
-                "type": "Planet Altitude",
-                "altitude": alt,
-            }
-            event_data["rarity"] = get_rarity("Planet Altitude", event_data)
-            event_data["duration"] = get_duration("Planet Altitude", event_data)
-            events.append(event_data)
-    logger.debug(f"--- calculate_highest_altitudes: {time.time() - start_time}s")
-    return events
-
-def calculate_aphelion_perihelion(start_date, end_date, executor):
-    start_time = time.time()
-    events = []
-    planets = planetary.APHELION_PERIHELION_PLANETS
-    futures = [
-        executor.submit(
-            skyfield_searches.find_aphelion_perihelion,
-            planet_name,
-            start_date,
-            end_date,
-        )
-        for planet_name in planets
-    ]
-    for future in as_completed(futures):
-        events_from_search = future.result()
-        for event_dict in events_from_search:
-            simple_name = planetary.get_simple_name(event_dict["planet"])
-            event_dict["event"] = event_dict["event_type"]
-            event_dict["object"] = simple_name
-            del event_dict["planet"]
-            del event_dict["event_type"]
-            event_dict["date"] = event_dict["date"].astimezone(utc)  # type: ignore
-            event_dict["type"] = "Aphelion/Perihelion"
-            event_dict["rarity"] = get_rarity(
-                "Aphelion/Perihelion", event_dict
-            )
-            event_dict["duration"] = get_duration("Aphelion/Perihelion", event_dict)
-            events.append(event_dict)
-    logger.debug(f"--- calculate_aphelion_perihelion: {time.time() - start_time}s")
-    return events
-
-def calculate_mercury_inferior_conjunctions(observer, start_date, end_date):
-    start_time = time.time()
-    events = skyfield_searches.find_mercury_inferior_conjunctions(
-        observer, start_date, end_date
-    )
-    for event in events:
-        event["type"] = "Inferior Conjunction"
-        if event.get("is_transit"):
-            event["event"] = "Mercury Transit"
-        else:
-            event["event"] = "Inferior Conjunction"
-        event["object"] = "Mercury"
-        event["rarity"] = get_rarity("Inferior Conjunction", event)
-        event["duration"] = get_duration("Inferior Conjunction", event)
-    logger.debug(
-        f"--- calculate_mercury_inferior_conjunctions: {time.time() - start_time}s"
-    )
-    return events
-
-def calculate_planet_alignments(observer, start_date, end_date):
-    start_time = time.time()
-    events = skyfield_searches.find_planet_alignments(
-        observer, start_date, end_date
-    )
-    for event in events:
-        event["type"] = "Planet Alignment"
-        event["rarity"] = get_rarity("Planet Alignment", event)
-        event["duration"] = get_duration("Planet Alignment", event)
-    logger.debug(f"--- calculate_planet_alignments: {time.time() - start_time}s")
-    return events
-
-def calculate_jovian_moon_events(observer, start_date, end_date):
-    start_time = time.time()
-    events = skyfield_searches.find_jovian_moon_events(
-        observer, start_date, end_date
-    )
-    for event in events:
-        event["rarity"] = get_rarity("Jovian Moon Event", event)
-        event["duration"] = get_duration("Jovian Moon Event", event)
-    logger.debug(f"--- calculate_jovian_moon_events: {time.time() - start_time}s")
-    return events
-
-def calculate_saturn_ring_crossings(start_date, end_date):
-    start_time = time.time()
-    events = skyfield_searches.find_saturn_ring_crossings(
-        start_date, end_date
-    )
-    for event in events:
-        event["date"] = event["date"].astimezone(utc)
-        event["rarity"] = get_rarity("Saturn Ring Crossing", event)
-        event["duration"] = get_duration("Saturn Ring Crossing", event)
-    logger.debug(
-        f"--- calculate_saturn_ring_crossings: {time.time() - start_time}s"
-    )
-    return events
-
-def calculate_greatest_elongations(observer, start_date, end_date):
-    start_time = time.time()
-    events = skyfield_searches.find_greatest_elongations(
-        observer, start_date, end_date
-    )
-    for event in events:
-        event["rarity"] = get_rarity("Greatest Elongation", event)
-        event["duration"] = get_duration("Greatest Elongation", event)
-    logger.debug(f"--- calculate_greatest_elongations: {time.time() - start_time}s")
-    return events
-
-def calculate_planetary_dichotomy(observer, start_date, end_date, precomputed_positions=None):
-    start_time = time.time()
-    events = skyfield_searches.find_planetary_dichotomy(
-        observer,
-        start_date,
-        end_date,
-        precomputed_positions=precomputed_positions,
-    )
-    for event in events:
-        event["rarity"] = get_rarity("Planetary Dichotomy", event)
-        event["duration"] = get_duration("Planetary Dichotomy", event)
-    logger.debug(f"--- calculate_planetary_dichotomy: {time.time() - start_time}s")
-    return events
-
-def calculate_planet_messier_conjunctions(observer, start_date, end_date, precomputed_positions=None):
-    start_time = time.time()
-    events = skyfield_searches.find_planet_messier_conjunctions(
-        observer,
-        start_date,
-        end_date,
-        precomputed_positions=precomputed_positions,
-    )
-    for event in events:
-        event["rarity"] = get_rarity("Planet-Messier Conjunction", event)
-        event["duration"] = get_duration("Planet-Messier Conjunction", event)
-    logger.debug(
-        f"--- calculate_planet_messier_conjunctions: {time.time() - start_time}s"
-    )
-    return events
-
-def calculate_jupiter_grs_transits(observer, start_date, end_date, grs_longitude=None):
-    start_time = time.time()
-    events = skyfield_searches.find_jupiter_grs_transits(
-        observer, start_date, end_date, grs_longitude=grs_longitude
-    )
-    for event in events:
-        event["rarity"] = get_rarity("Jupiter GRS Transit", event)
-        event["duration"] = get_duration("Jupiter GRS Transit", event)
-    logger.debug(f"--- calculate_jupiter_grs_transits: {time.time() - start_time}s")
-    return events
-
-def calculate_planet_star_conjunctions(observer, start_date, end_date, precomputed_positions=None):
-    start_time = time.time()
-    events = skyfield_searches.find_planet_star_conjunctions(
-        observer,
-        start_date,
-        end_date,
-        precomputed_positions=precomputed_positions,
-    )
-    for event in events:
-        event["rarity"] = get_rarity("Planet-Star Conjunction", event)
-        event["duration"] = get_duration("Planet-Star Conjunction", event)
-    logger.debug(
-        f"--- calculate_planet_star_conjunctions: {time.time() - start_time}s"
-    )
-    return events
-
-def calculate_planet_stationary_points(observer, start_date, end_date):
-    start_time = time.time()
-    events = []
-    planets = [
-        "mercury",
-        "venus",
-        "mars barycenter",
-        "jupiter barycenter",
-        "saturn barycenter",
-        "uranus barycenter",
-        "neptune barycenter",
-        "pluto barycenter",
-    ]
-    for p in planets:
-        found_events = skyfield_searches.find_stationary_points(
-            observer,
-            p,
-            start_date,
-            end_date,
-        )
-        for event in found_events:
-            event["rarity"] = get_rarity("Planet Stationary Point", event)
-            event["duration"] = get_duration("Planet Stationary Point", event)
-        events.extend(found_events)
-    logger.debug(
-        f"--- calculate_planet_stationary_points: {time.time() - start_time}s"
-    )
-    return events
-
-def calculate_planet_solar_conjunctions(observer, start_date, end_date):
-    start_time = time.time()
-    events = skyfield_searches.find_planet_solar_conjunctions(
-        observer, start_date, end_date
-    )
-    for event in events:
-        event["rarity"] = get_rarity("Planet Solar Conjunction", event)
-        event["duration"] = get_duration("Planet Solar Conjunction", event)
-    logger.debug(
-        f"--- calculate_planet_solar_conjunctions: {time.time() - start_time}s"
-    )
-    return events
-
-def calculate_planet_planet_occultations(observer, start_date, end_date):
-    start_time = time.time()
-    events = skyfield_searches.find_planet_planet_occultations(
-        observer, start_date, end_date
-    )
-    for event in events:
-        event["rarity"] = get_rarity("Planet-Planet Occultation", event)
-        event["duration"] = get_duration("Planet-Planet Occultation", event)
-    logger.debug(
-        f"--- calculate_planet_planet_occultations: {time.time() - start_time}s"
-    )
-    return events
-
-def calculate_mars_closest_approach(observer, start_date, end_date):
-    start_time = time.time()
-    events = skyfield_searches.find_mars_closest_approach(
-        start_date, end_date, observer=observer
-    )
-    for event in events:
-        event["rarity"] = get_rarity("Mars Closest Approach", event)
-        event["duration"] = get_duration("Mars Closest Approach", event)
-    logger.debug(f"--- calculate_mars_closest_approach: {time.time() - start_time}s")
-    return events
-
-def calculate_jovian_mutual_events(observer, start_date, end_date):
-    start_time = time.time()
-    events = skyfield_searches.find_jovian_mutual_events(
-        observer, start_date, end_date
-    )
-    for event in events:
-        event["rarity"] = get_rarity(event["type"], event)
-        event["duration"] = get_duration(event["type"], event)
-    logger.debug(f"--- calculate_jovian_mutual_events: {time.time() - start_time}s")
-    return events
